refactor(rekuho): refactor module and component structure
