+++ rekuho/src/lib/common.ts
