+++ rekuho/src/lib/linearAlgebra.ts
